Only parse code for Vector3 lib on demand
